refactor status
