+++ src/stores/logsStore.ts
@@ -1,6 +1,7 @@
 import { createStore, type Store } from "solid-js/store";
-import { buildStatusLog, type LogEntry } from "../utils/status.ts";
+import { isConnected, isConnecting, isDisconnected, isDisconnecting } from "../utils/status.ts";
 import { type StatusResponse } from "../services/vpnService.ts";
+import { formatDestination } from "../utils/destinations.ts";
 
 interface LogsState {
   logs: LogEntry[];
@@ -14,26 +15,77 @@ type LogsActions = {
 
 type LogsStoreTuple = readonly [Store<LogsState>, LogsActions];
 
+export type LogEntry = { date: string; message: string };
+
 export function createLogsStore(): LogsStoreTuple {
   const [state, setState] = createStore<LogsState>({ logs: [] });
 
-  const append = (message: string) => {
-    setState("logs", existing => {
-      const lastMessage = existing.length ? existing[existing.length - 1].message : "";
-      if (lastMessage === message) return existing;
-      const entry: LogEntry = { date: new Date().toISOString(), message };
-      return [...existing, entry];
-    });
-  };
+  function buildStatusLog(args: { response?: StatusResponse; error?: string }): string | undefined {
+    const lastMessage = state.logs.length ? state.logs[state.logs.length - 1].message : undefined;
+    return buildLogContent(args, lastMessage);
+  }
+
+  function buildLogContent(
+    args: {
+      response?: StatusResponse;
+      error?: string;
+    },
+    lastMessage?: string,
+  ): string | undefined {
+    let content: string | undefined;
+    if (args.response) {
+      const statusValue = args.response.status;
+      if (isConnected(statusValue)) {
+        const destination = statusValue.Connected;
+        const where = formatDestination(destination);
+        content = `Connected: ${where} - ${destination.address}`;
+      } else if (isConnecting(statusValue)) {
+        const destination = statusValue.Connecting;
+        const where = formatDestination(destination);
+        content = `Connecting: ${where} - ${destination.address}`;
+      } else if (isDisconnected(statusValue)) {
+        const lastWasDisconnected = Boolean(lastMessage && lastMessage.startsWith("Disconnected"));
+        if (lastWasDisconnected) {
+          content = undefined;
+        } else {
+          const lines = args.response.available_destinations.map(d => {
+            const where = formatDestination(d);
+            return `- ${where} - ${d.address}`;
+          });
+          content = `Disconnected. Available:\n${lines.join("\n")}`;
+        }
+      } else if (isDisconnecting(statusValue)) {
+        const destination = statusValue.Disconnecting;
+        const where = formatDestination(destination);
+        content = `Disconnecting: ${where} - ${destination.address}`;
+      } else {
+        const statusLabel = typeof statusValue === "string" ? statusValue : Object.keys(statusValue)[0] || "Unknown";
+        const destinations = args.response.available_destinations.length;
+        content = `status: ${statusLabel}, destinations: ${destinations}`;
+      }
+    } else if (args.error) {
+      content = `${args.error}`;
+    }
+    return content;
+  }
 
   const actions = {
-    append,
-    appendStatus: (response: StatusResponse) => {
-      const maybe = buildStatusLog(state.logs, { response });
-      if (maybe) append(maybe);
+    append: (message: string) => {
+      setState("logs", existing => {
+        const lastMessage = existing.length ? existing[existing.length - 1].message : "";
+        if (lastMessage === message) return existing;
+        const entry: LogEntry = { date: new Date().toISOString(), message };
+        return [...existing, entry];
+      });
     },
+
+    appendStatus: (response: StatusResponse) => {
+      const maybe = buildStatusLog({ response });
+      if (maybe) actions.append(maybe);
+    },
+
     clear: () => setState("logs", []),
-  } as const;
+  };
 
   return [state, actions] as const;
 }
